chore(app): rename typings
