+++ src/sc/index.d.ts
@@ -1,6 +1,6 @@
 export * from '../utils'
 export * from './typings/ContractParam';
 export * from './typings/core';
-export * from './typings/fDeserialize';
+export * from './typings/deserialize';
 export * from './typings/OpCode';
 export * from './typings/ScriptBuilder';
